Add bundle entrypoint
